external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-include.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error loading shader '[object WebGLShader]':INVALID_VALUE
vertex shader uses #include should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: vertex shader uses #include should fail
test vertex shader
default fragment shaderPASS vertex shader uses #include should fail
PASS successfullyParsed is true

TEST COMPLETE
